fixup! Add functionality for converting UNIX paths in arguments and environment variables to Windows form for native Win32 applications.
Let's prevent scp-style arguments from being mangled by MSYS2's path
conversion.

An argument like `me@example.com:/tmp/` is not something we should convert
into a Windows path; Use the absence of a slash before the colon as a
tell-tale that it is *not* a POSIX path list (exception: if the part
left of the colon is `.` or `..`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>